+++ src/client/index.js
@@ -1,3 +1,5 @@
+/* @flow */
+
 import React from 'react';
 import { render } from 'react-dom';
 import Router from 'react-router/lib/Router';