+++ src/server/htmlPage/render.js
@@ -1,11 +1,12 @@
+/* @flow */
 
 import { renderToString } from 'react-dom/server';
 import serialize from 'serialize-javascript';
 import Helmet from 'react-helmet';
 import clientAssets from '../clientAssets';
+import favicon from './favicon.ico';
 
-// :: [String] -> [String]
-function styleTags(styles) {
+function styleTags(styles : Array<string>) {
   return styles
     .map(style =>
       `<link href="${style}" media="screen, projection" rel="stylesheet" type="text/css" />`
@@ -13,8 +14,7 @@ function styleTags(styles) {
     .join('\n');
 }
 
-// :: [String] -> [String]
-function scriptTags(scripts) {
+function scriptTags(scripts : Array<string>) {
   return scripts
     .map(script =>
       `<script type="text/javascript" src="${script}"></script>`
@@ -37,7 +37,7 @@ const scripts = scriptTags(clientAssets.scripts);
  *
  * @return The full HTML page in the form of a React element.
  */
-function render(rootReactElement, initialState) {
+function render(rootReactElement : ?$React$Element, initialState : ?Object) {
   const reactRenderString = rootReactElement
     ? renderToString(rootReactElement)
     : null;
@@ -53,22 +53,22 @@ function render(rootReactElement, initialState) {
     : null;
 
   return `<!DOCTYPE html>
-    <html ${helmet && helmet.htmlAttributes.toString()}>
+    <html ${helmet ? helmet.htmlAttributes.toString() : ''}>
       <head>
         <meta charSet='utf-8' />
         <meta httpEquiv='X-UA-Compatible' content='IE=edge' />
         <meta httpEquiv='Content-Language' content='en' />
-        <link rel='shortcut icon' type='image/x-icon' href='/public/favicon.ico' />
+        <link rel='shortcut icon' type='image/x-icon' href='${favicon}' />
 
-        ${helmet && helmet.title.toString()}
-        ${helmet && helmet.meta.toString()}
-        ${helmet && helmet.link.toString()}
-        ${helmet && helmet.style.toString()}
+        ${helmet ? helmet.title.toString() : ''}
+        ${helmet ? helmet.meta.toString() : ''}
+        ${helmet ? helmet.link.toString() : ''}
 
         ${styles}
+        ${helmet ? helmet.style.toString() : ''}
       </head>
       <body>
-        <div id='app'>${reactRenderString}</div>
+        <div id='app'>${reactRenderString || ''}</div>
 
         <script type='text/javascript'>${
           initialState
@@ -76,8 +76,8 @@ function render(rootReactElement, initialState) {
             : ''
         }</script>
 
-        ${helmet && helmet.script.toString()}
         ${scripts}
+        ${helmet ? helmet.script.toString() : ''}
       </body>
     </html>`;
 }