+++ src/shared/components/Home/Home.js
@@ -1,3 +1,5 @@
+/* @flow */
+
 import React from 'react';
 import Helmet from 'react-helmet';
 
@@ -7,11 +9,12 @@ function Home() {
       <Helmet title="Home" />
 
       <p>
-        This is an alternative version of react-universally, a boilerplate that
-        contains an absolutely minimal set of dependencies in order to get you
-        up and running with a universal react project as quickly as possible.
-        It provides you with a great development experience that includes hot
-        reloading of everything.
+        This is an alternative version of react-universally (a starter kit that
+        contains all the build tooling and configuration you need to kick off
+        your next universal react project, whilst containing a minimal "project"
+        set up allowing you to make your own architecture decisions).  Please
+        reference the react-universally documentation for the features contained
+        within this starter kit.
       </p>
     </article>
   );